Drop equipped fitments and vehicles on death.
This changes the rules for death drops so that equipped fitments and
vehicles are (sometimes) dropped in addition to just inventories:

For characters that are killed directly in combat on the map, the
vehicle is still always destroyed; but with this change, any equipped
fitment has a 20% chance to survive and be dropped in addition to the
inventory.

For characters inside buildings that are destroyed, all fitments and
the vehicle itself are added to the total building inventory that then
has a 30% chance of being dropped.

This fixes https://github.com/xaya/taurion_gsp/issues/164.