Sort TargetKey same as targets/fighters.
Explicitly make the sorting order of TargetKey instances match the
order in which some database queries (TargetFinder::ProcessL1Targets and
FighterTable::ProcessWithAttacks) return targets, namely first buildings
and then characters (and within each group by ID).

For now, the sorting order of TargetKey is irrelevant, as it is just used
for direct lookups.  But in the future, we can then use it to sort targets
for processing, matching the direct DB queries (e.g. for parallelisation
of some things in combat).