Split computing from updating in target finding.
This refactors the target finding logic such that most of the computation
part is done without mutating anything, and the actual update to the
database is then done as a follow-up step based on the computed data.

For now this is just a refactor (including making things const that can
be now), but in the future, we can use this for parallelisation (do the
computation in parallel, and the updating on a single thread afterwards).