Return order data in game-state JSON.
This extends the game-state JSON to return data related to open orders
on the DEX inside buildings.  In particular, this includes three pieces
of information:

1) The JSON data for accounts (e.g. "getaccounts") has been extended to
include not just the available Cubit balance, but also the amount of
reserved Cubits (from open bids in any building) and the total (available
plus reserved).

2) The building JSON now returns "inventories" of reserved item quantities
from open asks in the new "reserved" field.  It has the same structure
as "inventories", except returning reserved instead of available balances.

3) The building JSON has a new field "orderbook", which returns all open
DEX orders inside that building.  They are grouped by item and structured
as an orderbook (split into bids and asks, and sorted by "best order first"
in each category).