Move quantity parsing to JSON util function.
Move the parsing and range-validation of an item quantity from the
move processor out into its own jsonutils function (so it can be reused
in other parts of the code).

We can also use the new function for parsing quantities in the
services JSON parser (which was previously using just its own
direct JSON parser).

Note that this also introduces a fork in consensus, namely by explicitly
forbidding values like 1.0 or 10e2 from being valid quantities.