Implement parsing and logic for transfer moves.
Set up the main logic (including JSON parsing) for handling
item transfer moves inside a building.  This also includes a
generic framework that can be extended for other DEX operations
in the future (like placing or cancelling orders).

For now we just have the code implemented and unit tests for it,
but it is not yet integrated with the move or pending processor.