Configuration of DEX fees.
Each trade on the DEX will cost a certain fee paid by the seller from the
Cubits they receive.  There will be a 3% base fee that is configured in
the ro params and just burned (to protect against spam and wash trading).

In addition to that, the owner of each building can configure a building
DEX fee in basis points up to 30% max, which is paid to the building owner
by the seller of each successful trade inside the building.

The building DEX fee is returned as a new field "dexfee" in the building
state JSON, and can be set with a new building-update move (similar to
setting the service fee):

  {"b": {"id": 42, "xf": 350}}

This would set the dex fee in building 42 to 350 bps, i.e. 3.5%.