Placing of bid/ask orders.
This implements new types of DEX operations, for placing bid or ask
limit orders.  This also implements order matching, which will happen
if a newly placed order can be filled already from existing ones.  If that
happens, then the remainder (if any) will be placed as limit order still.

To place a new bid or ask, the move format looks like this:

  {"x": [{"b": 100, "i": "foo", "n": 20, "bp": 5}]}
  {"x": [{"b": 100, "i": "foo", "n": 20, "ap": 100}]}