Enforce strict JSON integers.
When parsing move data, strictly enforce that values that should be
integers actually are of integer type (and not just happen to have
an integral value).  In particular, we do not (any more) allow things
like "1.0" or "2e2" to be in place of integers.

Some places in the code already had this check, but with this commit,
we extend it to all other places (that previously were just checking
for isInt or isUInt).

Fixes https://github.com/xaya/taurion_gsp/issues/186.