Restrict building construction by faction.
This associates a faction to some building types, based on the name
prefix of the type (i.e. "r vb" is red, while "huesli" has no faction).
This is done through RoConfig magic, similar to how bpo or prize items
are constructed.

Further, if a building has such a faction, it can only be constructed
(founded) by a character of the same faction.  In other words, a Reubo
can no longer found a Jodon vehicle bay, for instance.