Define factions for fitments/vehicles.
Add a new field to the ItemData roconfig proto, which can specify
the faction for faction-specific fitments and vehicles.  For fitments,
it is explicitly set in the roconfig data as appropriate.  For vehicles,
the faction will be deduced at runtime from the name prefix, e.g. "rv st".

In the future, these factions will be used to apply certain restrictions
(e.g. Jodon fitments can only be placed on Jodon vehicles).