Support arenas in LazyProto.
Add optional support to using an Arena to LazyProto.  If an arena
is set, then the message will be allocated on there when parsed or
SetToDefault.  If no arena is set, the message will be allocated
on the heap and managed by the LazyProto instance instead (as per
the previous behaviour).

For now, only the unit tests set an arena.  But we can use this to
employ an Arena per Database instance in the future and improve
memory performance.