Use arenas for database-extracted protos.
When extracting a protocol buffer message through the Database class,
use an arena for its allocation.  The arena is part of the Database
instance (which effectively means that one Arena is used per processed
block).

With this, memory allocation and deallocation overhead for parsing,
modifying and updating the protocol buffers in the database will be
reduced significantly.