Sub-message for building owner config.
Move the owner-configurable values of a building into a new and
dedicated sub-message (service fee and DEX trading fee).

The new sub-message is also reflected in the building game-state JSON.
Instead of having "dexfee" and "servicefee" directly in the building
JSON, they are now moved into a "config" object.

This will allow us to implement a general mechanism for building
updates (and delays for it with an ongoing-operation), independent
of the particular config fields updated / supported.