Ongoing operation for building config updates.
This defines a new ongoing operation for delayed updates to building
configuration.  When it "expires" (gets processed at its set block height),
the owner-configurable data in the associated building is updated from
some value set inside the operation.

In the future, when a building owner issues a move to update the building
configuration, we will schedule such an update rather than change the
config immediately.  This protects users in the building from frontrunning.