Implement UniqueHandles tracker.
This adds a new helper class, UniqueHandles, and an instance of it
to the Database.  With this class, we can track what DB handle instances
exist currently, to make sure that we are not accidentally having e.g.
two handles for the same account or character in existence at the same
time (which could lead to weird interactions / bugs).

For now, this just implements the basic infrastructure, and does not
actually enable tracking for any of the concrete handlers we have.