Database logic for account XP levels.
This adds the database schema and abstractions necessary to store
XPs per account and skill in the database.  This is handled with a new
helper class SkillManager, which is from the outside accessed as part
of the Account database handle.

For now, it simply reads and writes generic XP counts per skill; those
can be mapped to levels with some scheme later on.