Proto and config data for skills.
For the upcoming feature of skills, XP and character progression, this
adds some basic protocol buffer config data.  In particular, we define
an enum for skill types, and for each type, a config proto that can be
looked up through RoConfig.

For now, the config data only contains the skill's name, to be used
with game-state JSON as field names.  In the future, we might have
skill-specific constants, e.g. defining how that skill maps raw XP count
to "level".