Explicitly set CA file for Charon.
Charon recently enabled strict checks for the XMPP server's certificate,
https://github.com/xaya/charon/pull/16.  With this change, we add a new
--charon_cafile argument, which can be used to explicitly set the
trust root for that when using the Charon client or server built into
Taurion.  To avoid confusion, the existing --cafile argument (for the
REST client) is renamed to --rest_cafile.

We also update the Charon integration test (charon.py) to explicitly
set the CA file now, based on the one shipped and installed with
Charon in /usr/local/share/charon.