Try to avoid turns in path stepping.
When stepping paths, try to continue the last direction as long
as possible while still retaining the optimal distance.  This helps
to avoid unnecessary turns (which then lead to unnecessary many waypoints
in moves).

Note that the implementation is very simple and just a greedy approach, so
it is not (?) guaranteed to really minimise the number of turns.  It does so
quite well in practice, though.

Since the path finder is not part of consensus, this change is not forking
the game state (nor would further tweaks to the algorithm cause forks).