Handle burn in pxtest.py.
Upstream xayagametest has removed the burn argument to sendMove,
as it is specific to Xaya Core while xayagametest is agnostic to
the underlying blockchain environment.

Burns are only really used in Taurion anyway for now, so we can just
handle the simple logic for setting burns in the Xaya Core options
there directly.