Enable burns on Polygon.
Enable burns on Polygon (for minting credits).  Burning is not directly
supported by the EVM Xaya framework (especially on Polygon), as we do not
want the "burnt" coins to remain sitting in the Polygon bridge forever.

Instead, "burning" for the purpose of the game is done by sending to another
special address.  This burn address is also controlled by the Xaya team, but
with the understanding that coins received there will be bridged back to
Ethereum and burnt there from time to time.